+++ muikku-atests/src/test/java/fi/muikku/AbstractIntegrationTest.java
@@ -14,6 +14,11 @@ import java.util.regex.Pattern;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.http.client.methods.HttpPost;
+import org.apache.http.entity.StringEntity;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClients;
+import org.apache.http.util.EntityUtils;
 import org.joda.time.DateTime;
 import org.junit.After;
 import org.junit.Before;
@@ -106,6 +111,29 @@ public abstract class AbstractIntegrationTest extends TestWatcher {
     }
   }
 
+  protected Boolean webhookCall(String url, String payload) throws Exception {
+    String signature = "38c6cbd28bf165070d070980dd1fb595";
+    CloseableHttpClient client = HttpClients.createDefault();
+    try {
+      HttpPost httpPost = new HttpPost(url);
+      try {
+        StringEntity dataEntity = new StringEntity(payload);
+        try {
+          httpPost.addHeader("X-Pyramus-Signature", signature);
+          httpPost.setEntity(dataEntity);
+          client.execute(httpPost);
+          return true;
+        } finally {
+          EntityUtils.consume(dataEntity);
+        }
+      } finally {
+        httpPost.releaseConnection();
+      }
+    } finally {
+      client.close();
+    }
+  }
+  
   private Connection getConnection() throws SQLException, ClassNotFoundException {
     Class.forName(getJdbcDriver());
     return DriverManager.getConnection(getJdbcUrl(), getJdbcUsername(), getJdbcPassword());
